+++ tests/integration/test_tls.py
@@ -20,10 +20,30 @@ METADATA = yaml.safe_load(Path("./charmcraft.yaml").read_text())
 APP_NAME = "tempo"
 SSC = "self-signed-certificates"
 SSC_APP_NAME = "ssc"
+TRAEFIK = "traefik-k8s"
+TRAEFIK_APP_NAME = "trfk"
 TRACEGEN_SCRIPT_PATH = Path() / "scripts" / "tracegen.py"
 
 
 logger = logging.getLogger(__name__)
+
+
+async def get_ingress_proxied_hostname(ops_test: OpsTest):
+    status = await ops_test.model.get_status()
+    app = status["applications"][TRAEFIK_APP_NAME]
+    status_msg = app["status"]["info"]
+
+    # hacky way to get ingress hostname
+    if "Serving at" not in status_msg:
+        assert False, f"Ingressed hostname is not present in {TRAEFIK_APP_NAME} status message."
+    return status_msg.replace("Serving at", "").strip()
+
+
+async def get_tempo_ingressed_endpoint(hostname, protocol):
+    protocol_endpoint = protocols_endpoints.get(protocol)
+    if protocol_endpoint is None:
+        assert False, f"Invalid {protocol}"
+    return protocol_endpoint.format(hostname)
 
 
 async def get_tempo_traces_internal_endpoint(ops_test: OpsTest, protocol):
@@ -59,8 +79,7 @@ async def test_build_and_deploy(ops_test: OpsTest):
             apps=[APP_NAME, SSC_APP_NAME],
             status="active",
             raise_on_blocked=True,
-            timeout=10000,
-            raise_on_error=False,
+            timeout=1000,
         ),
     )
 
@@ -99,25 +118,8 @@ async def test_verify_trace_http_no_tls_fails(ops_test: OpsTest, server_cert, no
 
 
 @pytest.mark.abort_on_fail
-@pytest.mark.parametrize("protocol", list(protocols_endpoints.keys()))
-async def test_verify_traces_force_enabled_protocols_tls(ops_test: OpsTest, nonce, protocol):
-
-    tempo_app: Application = ops_test.model.applications[APP_NAME]
-
-    # enable each protocol receiver
-    # otlp_http should be enabled by default
-    if protocol != "otlp_http":
-        await tempo_app.set_config(
-            {
-                f"always_enable_{protocol}": "True",
-            }
-        )
-        await ops_test.model.wait_for_idle(
-            apps=[APP_NAME, WORKER_NAME],
-            status="active",
-            timeout=1000,
-        )
-
+async def test_verify_traces_otlp_http_tls(ops_test: OpsTest, nonce):
+    protocol = "otlp_http"
     tempo_endpoint = await get_tempo_traces_internal_endpoint(ops_test, protocol=protocol)
     # WHEN we emit a trace secured with TLS
     await emit_trace(
@@ -127,6 +129,67 @@ async def test_verify_traces_force_enabled_protocols_tls(ops_test: OpsTest, nonc
     await get_traces_patiently(
         await get_application_ip(ops_test, APP_NAME), service_name=f"tracegen-{protocol}"
     )
+
+
+@pytest.mark.abort_on_fail
+async def test_deploy_ingress(ops_test: OpsTest):
+    await asyncio.gather(
+        ops_test.model.deploy(
+            TRAEFIK, application_name=TRAEFIK_APP_NAME, channel="edge", trust=True
+        ),
+    )
+    await asyncio.gather(
+        ops_test.model.wait_for_idle(
+            apps=[TRAEFIK_APP_NAME],
+            status="active",
+            raise_on_blocked=True,
+            timeout=600,
+        ),
+    )
+
+
+@pytest.mark.abort_on_fail
+async def test_relate_ingress(ops_test: OpsTest):
+    await ops_test.model.integrate(
+        SSC_APP_NAME + ":certificates", TRAEFIK_APP_NAME + ":certificates"
+    )
+    await ops_test.model.integrate(APP_NAME + ":ingress", TRAEFIK_APP_NAME + ":traefik-route")
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, SSC_APP_NAME, TRAEFIK_APP_NAME, WORKER_NAME],
+        status="active",
+        timeout=1000,
+    )
+
+
+@pytest.mark.abort_on_fail
+async def test_force_enable_protocols(ops_test: OpsTest):
+    tempo_app: Application = ops_test.model.applications[APP_NAME]
+    config = {}
+    for protocol in list(protocols_endpoints.keys()):
+        if protocol != "otlp_http":
+            config[f"always_enable_{protocol}"] = "True"
+
+    await tempo_app.set_config(config)
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, WORKER_NAME],
+        status="active",
+        timeout=1000,
+    )
+
+
+@pytest.mark.abort_on_fail
+async def test_verify_traces_force_enabled_protocols_tls(ops_test: OpsTest, nonce):
+
+    tempo_host = await get_ingress_proxied_hostname(ops_test)
+
+    for protocol in list(protocols_endpoints.keys()):
+        tempo_endpoint = await get_tempo_ingressed_endpoint(tempo_host, protocol=protocol)
+        # emit a trace secured with TLS
+        await emit_trace(
+            tempo_endpoint, ops_test, nonce=nonce, verbose=1, proto=protocol, use_cert=True
+        )
+        # verify it's been ingested
+        await get_traces_patiently(tempo_host, service_name=f"tracegen-{protocol}")
 
 
 @pytest.mark.teardown
